processing/requirements.txt
modal>=0.64.0
deepgram-sdk>=3.0.0
google-generativeai>=0.8.0
firebase-admin>=6.0.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
pydantic>=2.0.0
mediapipe>=0.10.0
